improve/refactor test case structure
