+++ demo/demo-pojo/pojo-server/src/main/java/org/apache/servicecomb/demo/pojo/server/TestImpl.java
@@ -105,4 +105,9 @@ public class TestImpl implements Test {
   public String testTraceId() {
     return ContextUtils.getInvocationContext().getContext(CoreConst.TRACE_ID_NAME);
   }
+
+  @Override
+  public int[] testIntArray(int[] request) {
+    return request;
+  }
 }
