+++ demo/demo-schema/src/main/java/org/apache/servicecomb/demo/server/Test.java
@@ -35,4 +35,6 @@ public interface Test {
   String addString(String[] strArr);
 
   String testTraceId();
+
+  int[] testIntArray(int[] request);
 }
